Fix incorrect Delta Metric calculation
Problem:
Delta metrics were not calculated properly because the only the delta value of network stats was persisted in a metrics object.

Solution:
Introduce a second field to store the aggregate network stats, and use that value in delta calculation

Testing:
Introduce a 3rd metric to the delta metrics test to ensure that the delta is not used in subsequent calculations.